Created util package in kangaroo-common
Moved getOne from AbstractAuthenticator into new ParamUtil.
Moved PasswordUtil from security into util package.
