+++ kangaroo-common/src/test/java/net/krotscheck/kangaroo/util/PasswordUtilTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package net.krotscheck.kangaroo.common.security;
+package net.krotscheck.kangaroo.util;
 
 import org.apache.commons.lang3.RandomStringUtils;
 import org.junit.Assert;
