+++ kangaroo-server/src/test/java/net/krotscheck/test/EnvironmentBuilder.java
@@ -28,8 +28,8 @@ import net.krotscheck.kangaroo.database.entity.OAuthTokenType;
 import net.krotscheck.kangaroo.database.entity.Role;
 import net.krotscheck.kangaroo.database.entity.User;
 import net.krotscheck.kangaroo.database.entity.UserIdentity;
-import net.krotscheck.kangaroo.common.security.PasswordUtil;
 import net.krotscheck.kangaroo.test.IFixture;
+import net.krotscheck.kangaroo.util.PasswordUtil;
 import org.apache.commons.lang3.StringUtils;
 import org.hibernate.Session;
 import org.hibernate.Transaction;
